scripts/requirements.txt
# PixelLab Gallery Generator Dependencies

# HTTP client for API calls
requests>=2.31.0

# Environment variable management
python-dotenv>=1.0.0

# Image processing (optional, for advanced manipulation)
Pillow>=10.0.0
